Add email tips to Environment.{define,assign}
> Me and my partner are having trouble getting started with the project.
> We have started to try to implement the environment class, but cannot
> seem to make any headway.  What exactly are the functions define and
> assign supposed to be doing?  Where in the Environment are we

The association list is for the local table, i.e., field "frame" in
class Environment.  The method assign() is for implementing Set.eval().
It modifies the value of an existing variable.  The recursive structure
is the same as for lookup(), except that after calling find(), the value
should be updated instead of returned.  The method define() is for
Define.eval().  It only adds a new definition to the current frame,
without following the chain of Environment objects.